Extract file upload component for general s3 upload usage
