+++ client/components/common/FileUpload.vue
@@ -1,30 +1,35 @@
 <template>
-  <div class="file-upload">
-    <div class="upload-button">
-      <label for="upload" class="upload-label btn btn-link btn-sm">
-        <span class="mdi mdi-file"></span> Upload
-        <input @change="upload" id="upload" type="file" class="upload-input"/>
-      </label>
-    </div>
+  <div class="file-upload-form">
+    <form action="http://sigv4examplebucket.s3.amazonaws.com/" method="post" enctype="multipart/form-data">
+      Key to upload:
+      <input type="input" name="key" value="user/user1/${filename}" /><br />
+      <input type="hidden" name="acl" value="public-read" />
+      <input type="hidden" name="success_action_redirect" value="http://sigv4examplebucket.s3.amazonaws.com/successful_upload.html" />
+      Content-Type:
+      <input type="input" name="Content-Type" value="image/jpeg" /><br />
+      <input type="hidden" name="x-amz-meta-uuid" value="14365123651274" />
+      <input type="hidden" name="x-amz-server-side-encryption" value="AES256" />
+      <input type="text" name="X-Amz-Credential" value="AKIAIOSFODNN7EXAMPLE/20151229/us-east-1/s3/aws4_request" />
+      <input type="text" name="X-Amz-Algorithm" value="AWS4-HMAC-SHA256" />
+      <input type="text" name="X-Amz-Date" value="20151229T000000Z" />
+      Tags for File:
+      <input type="input" name="x-amz-meta-tag" value="" /><br />
+      <input type="hidden" name="Policy" value="<Base64-encoded policy string>" />
+      <input type="hidden" name="X-Amz-Signature" value="<signature-value>" />
+      File:
+      <input type="file" name="file" /> <br />
+      <!-- The elements after this will be ignored -->
+      <input type="submit" name="submit" value="Upload to Amazon S3" />
+    </form>
   </div>
 </template>
 
 <script>
-import isEmpty from 'lodash/isEmpty';
+// import request from '../../api/request';
 
 export default {
   name: 'file-upload',
-  methods: {
-    upload({ target }) {
-      const file = !isEmpty(target.files) ? target.files[0] : null;
-      if (!file) return;
-      const reader = new window.FileReader();
-      reader.readAsDataURL(file);
-      reader.addEventListener('load', e => {
-        window.alert('uploaded');
-      });
-    }
-  }
+  methods: {}
 };
 </script>
