+++ client/components/common/Meta/index.vue
@@ -10,7 +10,7 @@
 import Checkbox from './Checkbox';
 import ColorPicker from './ColorPicker';
 import DatePicker from './DatePicker';
-import FileUpload from './FileUpload';
+import FileUpload from './File';
 import Input from './Input';
 import mapKeys from 'lodash/mapKeys';
 import Select from './Select';
